// frontend/src/app/api/model/milestone.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Milestone { 
    id: number;
    cardId: number;
    title: string;
    done: boolean;
    targetDate?: string | null;
}